Moved non-belonging things from GLSLElementTypes
